fix: Fix norm issues under utils folder
