+++ source/debug/show_list.c
@@ -13,6 +13,19 @@
 #include "minishell.h"
 #include "parser.h"
 
+void	print_token(t_token *token)
+{
+	printf("(type: %s, data: %s)\n",
+		ft_get_token_type_str(token->type), token->data);
+}
+
+void	print_token_list(t_list *token_list)
+{
+	printf("-------- TOKEN LIST --------\n");
+	ft_lstiter(token_list, (void *)print_token);
+	printf("----------------------------\n\n");
+}
+
 void	ft_print_env(t_env *env)
 {
 	printf("%s=%s\n", env->key, env->value);
