+++ source/utils/stack_utils.c
@@ -62,34 +62,3 @@ t_token	*get_token_from_stack(t_list *node)
 		return (NULL);
 	return ((t_token *)node->content);
 }
-
-void	print_state_stack(t_list *stack)
-{
-	t_list	*node;
-
-	node = stack;
-	while (node)
-	{
-		printf("%d <- ", get_state_from_stack(node));
-		node = node->next;
-	}
-	printf("(NULL)\n");
-}
-
-t_ast	*get_ast_from_stack(t_list *node)
-{
-	if (!node)
-		return (NULL);
-	return ((t_ast *)node->content);
-}
-
-void	print_parse_stack(t_list *node)
-{
-	while (node)
-	{
-		if (get_token_from_stack(node))
-			printf("%d <- ", get_ast_from_stack(node)->type);
-		node = node->next;
-	}
-	printf("(NULL)\n");
-}
