+++ source/utils/token_status_utils.c
@@ -1,24 +1,34 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   io_redirect_status_utils.c                         :+:      :+:    :+:   */
+/*   token_status_utils.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lyeh <lyeh@student.42vienna.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/02/17 23:49:36 by lyeh              #+#    #+#             */
-/*   Updated: 2024/02/17 23:49:38 by lyeh             ###   ########.fr       */
+/*   Created: 2024/02/17 23:53:21 by lyeh              #+#    #+#             */
+/*   Updated: 2024/02/17 23:53:22 by lyeh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "defines.h"
-#include "utils.h"
 
-int	get_redirect_type_from_list(t_list *io_red_list)
+t_token	*get_token_from_list(t_list *token_list)
 {
-	t_io_red	*io_red;
+	if (!token_list)
+		return (NULL);
+	return ((t_token *)token_list->content);
+}
 
-	if (!io_red_list)
+int	get_token_type_from_list(t_list *token_list)
+{
+	if (!token_list)
 		return (T_NONE);
-	io_red = io_red_list->content;
-	return (io_red->type);
+	return (((t_token *)token_list->content)->type);
+}
+
+char	*get_token_data_from_list(t_list *token_list)
+{
+	if (!token_list)
+		return (NULL);
+	return (((t_token *)token_list->content)->data);
 }
